+++ extensions/speedtest/CHANGELOG.md
@@ -5,6 +5,7 @@
 - Catch possible crash in speedtest
 - Upgrade to 1.57
 - Use Raycast Icons instead of custom ones
+- Add Restart Action
 
 ## [Spelling] - 2023-06-13